Session Service Validation Promises
Got the session service working with promises when the service needs to be
validated, this is used for proper page initialization, which drastically
helps the service. I know @julianpoy doesn't like promises, but I couldnt
find a better way to properly do this without using promises. Hope the code
looks great!